requirements/requirements.txt
appdirs==1.4.3
attrs==19.1.0
boto3==1.9.116
botocore==1.12.116
certifi==2019.3.9
codeclimate-test-reporter==0.2.3
coverage==4.3.4
dj-database-url==0.5.0
django-cors-headers==1.2.2
django-debug-toolbar==1.7
django-extensions==1.7.5
django-filter==0.14.0
git+https://github.com/fedspendingtransparency/django-mock-queries#egg=django-mock-queries
django-queryset-csv==1.0.1
django-simple-history==1.8.2
django-spaghetti-and-meatballs==0.2.2
Django==1.11.22
djangorestframework==3.9.4
drf-extensions==0.3.1
drf-tracking==1.4.0
django-partial-index==0.5.2
elasticsearch==6.2.0
fiscalyear==0.1.0
flake8==3.7.7
Markdown==2.6.6
model-mommy==1.6.0
numpy==1.9.0
openpyxl==2.4.7
pandas==0.23.3
psycopg2-binary==2.7.5
py-gfm==0.1.3
py==1.5.4
pytest-cov==2.5.1
pytest-django==3.2.1
pytest==3.10.1
python-dateutil==2.7.3
python-json-logger==0.1.9
pytz==2016.6.1
requests==2.20.0
retrying==1.3.3
six==1.10.0
unicodecsv==0.14.1
urllib3==1.24.2
xlrd==1.0.0
jsonpickle==0.9.6
filechunkio==1.8
pip==19.2.1
pyyaml==5.1.2
